+++ src/applications/accredited-representative-portal/components/Error/Error404.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+const Error404 = () => {
+  return (
+    <>
+      <h2 slot="headline">We’re sorry. Something went wrong.</h2>
+      <div>
+        <p className="vads-u-margin-y--0">
+          Please refresh this page or check back later. You can also check the
+          system status on the VA.gov homepage.
+        </p>
+      </div>
+    </>
+  );
+};
+
+export default Error404;
